+++ engine_src/shared/T1_material.c
@@ -7,7 +7,7 @@
 LockedMaterialCollection * all_mesh_materials = NULL;
 
 void T1_material_init(
-    void * (* arg_malloc_function)(u64 size))
+    void * (* arg_malloc_function)(size_t size))
 {
     all_mesh_materials = arg_malloc_function(sizeof(LockedMaterialCollection));
     all_mesh_materials->size = 0;
