+++ engine_src/shared/T1_meta.h
@@ -15,7 +15,7 @@
 
 void T1_meta_init(
     void *(* T1_meta_memcpy)(void *, const void *, u64),
-    void *(* malloc_func)(u64),
+    void *(* malloc_func)(size_t),
     void *(* memset_func)(void *, s32, u64),
     s32 (* strcmp_func)(const char *, const char *),
     u64 (* strlen_func)(const char *),
