+++ engine_src/shared/T1_settings.h
@@ -10,7 +10,7 @@ typedef struct {
 
 extern T1PerfSettings * T1_perf_settings;
 
-void T1_settings_init(void * arg_malloc_func(u64));
+void T1_settings_init(void * arg_malloc_func(size_t));
 
 u32 T1_settings_get_render_width(void);
 u32 T1_settings_get_render_height(void);
