+++ engine_src/shared/T1_token.c
@@ -100,7 +100,7 @@ static TokTokenState * tts = NULL;
 void T1_token_init(
     void * (* arg_memset_func)(void *, int, u64),
     u64 (* arg_strlen_func)(const char *),
-    void * (* arg_malloc_func)(u64),
+    void * (* arg_malloc_func)(size_t),
     u8 * good)
 {
     *good = 0;
